+++ SudoSodoku/Views/LandingView.swift
@@ -53,7 +53,7 @@ struct LandingView: View {
                             }
                         }
                     
-                    Text("KERNEL_V1.0.0").font(.system(size: 14, design: .monospaced)).foregroundColor(.gray).padding(.top, 5)
+                    Text("KERNEL_V\(AppConstants.marketingVersion)").font(.system(size: 14, design: .monospaced)).foregroundColor(.gray).padding(.top, 5)
                 }
                 Spacer()